Full CRUD

+++ src/main/java/Kon/repositories/ConsoleSalesRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ConsoleSalesRepository extends JpaRepository<ConsoleSalesModel, Long> {
+public interface ConsoleSalesRepository extends JpaRepository<ConsoleSalesModel, Integer> {
 
 }
